add preprocessing
